+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## v2.1.0 - 2022.08.12 10
+* support BigInt and format `q/Q` for 8 bytes number
+* improve `locale` keys and translations
+* bump up `@nuogz/i18n` to `1.2.0` and update related code
+
+
 ## v2.0.2 - 2022.08.09 08
 * change library's description
 * bump up `@nuogz/i18n` to `1.0.2`